Embed rewards and vector store into core chain
